move StateChange subclass to signals.py
